The MOUSE Update
MouseListener added
Bullets shoot at where clicked
Added grenades
Players now shoot grenade projectiles at a slower speed
Point, vector, and all mobs switched to double

+++ Gambit/src/com/lasacsgames/game/entity/mob/GhostEnemy.java
@@ -15,6 +15,6 @@ public class GhostEnemy extends Mob
 
 	public void render(Screen screen)
 	{
-		if (drawable) screen.renderEntity(location.x - 8, location.y - 8, Sprite.ghost);
+		if (drawable) screen.renderEntity((int)location.x - 8, (int)location.y - 8, Sprite.ghost);
 	}
 }
